refactor(architecture): reorganize domain layer into entities, repositories, value objects and enums
