+++ src/aigora/curriculum_graph/domain/__init__.py
@@ -1,17 +1,8 @@
-from .curriculum_graph import CurriculumGraph
-from .curriculum_profile import CurriculumProfile
-from .edge import Edge
-from .enums import EdgeType, MasteryLevel
-from .mastery import MasteryCriterion, MasteryScale
-from .node import Node
+from aigora.curriculum_graph.domain.entities.curriculum_graph import CurriculumGraph
+from aigora.curriculum_graph.domain.entities.curriculum_profile import CurriculumProfile
+from aigora.curriculum_graph.domain.entities.edge import Edge
+from aigora.curriculum_graph.domain.entities.node import Node
+from aigora.curriculum_graph.domain.enums.enums import EdgeType, MasteryLevel
+from aigora.curriculum_graph.domain.value_objects.mastery import MasteryCriterion, MasteryScale
 
-__all__ = [
-    "CurriculumGraph",
-    "CurriculumProfile",
-    "Edge",
-    "EdgeType",
-    "MasteryCriterion",
-    "MasteryLevel",
-    "MasteryScale",
-    "Node",
-]
+__all__ = ["CurriculumGraph", "CurriculumProfile", "Edge", "EdgeType", "MasteryCriterion", "MasteryLevel", "MasteryScale", "Node"]
