+++ src/aigora/curriculum_graph/domain/entities/__init__.py
@@ -1,0 +1,6 @@
+from aigora.curriculum_graph.domain.entities.curriculum_graph import CurriculumGraph
+from aigora.curriculum_graph.domain.entities.curriculum_profile import CurriculumProfile
+from aigora.curriculum_graph.domain.entities.edge import Edge
+from aigora.curriculum_graph.domain.entities.node import Node
+
+__all__ = ["CurriculumGraph", "CurriculumProfile", "Edge", "Node"]
