+++ src/aigora/curriculum_graph/domain/enums/__init___.py
@@ -1,0 +1,3 @@
+from aigora.curriculum_graph.domain.enums.enums import EdgeType, MasteryLevel
+
+__all__ = ["EdgeType", "MasteryLevel"]
