+++ src/aigora/curriculum_graph/domain/enums/enums.py
