+++ src/aigora/curriculum_graph/domain/repositories/__init__.py
@@ -1,0 +1,3 @@
+from aigora.curriculum_graph.domain.repositories.graph_repository import GraphRepository
+
+__all__ = ["GraphRepository"]
