+++ src/aigora/curriculum_graph/domain/value_objects/__init__.py
@@ -1,0 +1,3 @@
+from aigora.curriculum_graph.domain.value_objects.mastery import MasteryCriterion, MasteryScale
+
+__all__ = ["MasteryCriterion", "MasteryScale"]
